MANUAL: Show for which formats an extension is supported on hover
This commit introduces a new extension-support.lua Lua filter
for MANUAL.html that recognizes headings like the following:

#### Extension: `fenced_code_blocks` ####

And appends a Span with content "±" to the heading.
The Span has a title attribute (which browsers display on hover)
with automatically generated text like the following:

  enabled by default for:
  + ipynb
  + markdown
  + markdown_github
  + markdown_phpextra
  + opml

  disabled by default for:
  - markdown_mmd
  - markdown_strict
  - plain

Letting readers quickly check for which formats a specific extension
is supported / enabled by default / disabled by default.